requirements.txt
splunk-sdk
sigma-cli
python-dotenv